Extract db upgrade code from queries.ts into its own file.
